small core updates
