+++ src/modules/providers/pipes/validate-provider.pipe.ts
@@ -23,35 +23,35 @@ export class ValidateProviderPipe implements PipeTransform {
       if (Object.keys(value).length === 0) {
         throw new BadRequestException('At least a field is required to update');
       }
-      await this.whenUpdateOrDelete(value as UpdateProviderDto);
+      await this.#whenUpdateOrDelete(value as UpdateProviderDto);
     } else {
-      await this.whenCreate(value as CreateProviderDto);
+      await this.#whenCreate(value as CreateProviderDto);
     }
 
     return value;
   }
 
-  whenUpdateOrDelete = async (value: UpdateProviderDto) => {
+  #whenUpdateOrDelete = async (value: UpdateProviderDto) => {
     //Record found during update
     if (!isMissing(value) && typeof value === 'string') {
       const isRecrodFound = await this.providersService.findProviderByValue({ id: value });
       if (!isRecrodFound) throw new NotFoundException(`Record not found with id: ${value}`);
     } else {
       if (isMissing(value?.vault)) throw new BadRequestException('Vault id is required');
-      if (!isMissing(value?.name)) await this.checkProviderNameIsUnique(value);
+      if (!isMissing(value?.name)) await this.#checkProviderNameIsUnique(value);
     }
 
     return true;
   };
 
-  whenCreate = async (value: CreateProviderDto) => {
+  #whenCreate = async (value: CreateProviderDto) => {
     //Validate provider If it is unique by name or not
-    await this.checkProviderNameIsUnique(value);
+    await this.#checkProviderNameIsUnique(value);
 
     return true;
   };
 
-  checkProviderNameIsUnique = async (value?: { name?: string; vault?: string }) => {
+  #checkProviderNameIsUnique = async (value?: { name?: string; vault?: string }) => {
     const isNameUnique = await this.providersService.findProviderByValue({
       name: value?.name,
       vault: value?.vault,
